Distribute functions to new file/modules
Rewrite test file for import of functions from different files
